+++ src/formatters.rs
@@ -139,7 +139,7 @@ pub fn s2v_f32_hz_then_khz() -> Arc<dyn Fn(&str) -> Option<f32> + Send + Sync> {
     Arc::new(move |string| {
         // If the user inputs a note representation, then we'll use that
         if let Some(midi_note_number) = note_formatter(string) {
-            return Some(util::midi_note_to_freq(midi_note_number.clamp(0, 127) as u8) as f32);
+            return Some(util::midi_note_to_freq(midi_note_number.clamp(0, 127) as u8));
         }
 
         let string = string.trim();
